+++ packages/@expo/cli/src/start/interface/interactiveActions.ts
@@ -19,11 +19,21 @@ export class DevServerManagerActions {
   ) {
     // If native dev server is running, print its URL.
     if (this.devServerManager.getNativeDevServerPort()) {
+      const devServer = this.devServerManager.getDefaultDevServer();
       try {
-        const url = this.devServerManager.getDefaultDevServer().getNativeRuntimeUrl()!;
+        const nativeRuntimeUrl = devServer.getNativeRuntimeUrl()!;
+        const interstitialPageUrl = devServer.getRedirectUrl();
 
-        printQRCode(url);
-        Log.log(printItem(chalk`Metro waiting on {underline ${url}}`));
+        printQRCode(interstitialPageUrl ?? nativeRuntimeUrl);
+
+        if (interstitialPageUrl) {
+          Log.log(
+            printItem(
+              chalk`Choose an app to open your project at {underline ${interstitialPageUrl}}`
+            )
+          );
+        }
+        Log.log(printItem(chalk`Metro waiting on {underline ${nativeRuntimeUrl}}`));
         // TODO: if development build, change this message!
         Log.log(printItem('Scan the QR code above with Expo Go (Android) or the Camera app (iOS)'));
       } catch (error) {
@@ -31,7 +41,7 @@ export class DevServerManagerActions {
         if (error.code !== 'NO_DEV_CLIENT_SCHEME') {
           throw error;
         } else {
-          const serverUrl = this.devServerManager.getDefaultDevServer().getDevServerUrl();
+          const serverUrl = devServer.getDevServerUrl();
           Log.log(printItem(chalk`Metro waiting on {underline ${serverUrl}}`));
           Log.log(printItem(`Linking is disabled because the client scheme cannot be resolved.`));
         }