+++ packages/@expo/cli/src/start/platforms/ios/ApplePlatformManager.ts
@@ -12,8 +12,10 @@ export class ApplePlatformManager extends PlatformManager<Device> {
     options: {
       /** Get the base URL for the dev server hosting this platform manager. */
       getDevServerUrl: () => string | null;
-      /** Expo Go URL */
-      getExpoGoUrl: () => string | null;
+      /** Expo Go URL. */
+      getExpoGoUrl: () => string;
+      /** Get redirect URL for native disambiguation. */
+      getRedirectUrl: () => string | null;
       /** Dev Client */
       getCustomRuntimeUrl: (props?: { scheme?: string }) => string | null;
     }