+++ packages/expo-modules-core/android/src/main/cpp/types/ExpectedType.h
@@ -22,6 +22,8 @@ public:
   CppType getCppType();
 
   jni::local_ref<jni::JavaClass<ExpectedType>::javaobject> getFirstParameterType();
+
+  jni::local_ref<jni::JavaClass<ExpectedType>::javaobject> getSecondParameterType();
 };
 
 /**
@@ -41,5 +43,7 @@ public:
    * If the allowsPrimitives is set to true type like int will be represented as a primitives.
    */
   std::string getJClassString(bool allowsPrimitives = false);
+
+  jni::local_ref<jni::JArrayClass<SingleType>::javaobject> getPossibleTypes();
 };
 } // namespace expo